+++ src/App.tsx
@@ -7,7 +7,7 @@ import {
   ToasterProvider,
 } from "@chainsafe/common-components";
 
-import Routes from "./Components/Routes";
+import ChainbridgeRoutes from "./Components/Routes";
 import { lightTheme } from "./Themes/LightTheme";
 import { ChainbridgeProvider } from "./Contexts/ChainbridgeContext";
 import AppWrapper from "./Layouts/AppWrapper";
@@ -31,6 +31,12 @@ if (
 }
 
 const App: React.FC<{}> = () => {
+  const {
+    __RUNTIME_CONFIG__: {
+      UI: { wrapTokenPage = false } = {},
+      CHAINBRIDGE: { chains },
+    },
+  } = window;
   const tokens = chainbridgeConfig.chains
     .filter((c) => c.type === "Ethereum")
     .reduce((tca, bc: any) => {
@@ -87,10 +93,10 @@ const App: React.FC<{}> = () => {
             gasPriceSetting="fast"
           >
             <NetworkManagerProvider>
-              <ChainbridgeProvider>
+              <ChainbridgeProvider chains={chains}>
                 <Router>
-                  <AppWrapper>
-                    <Routes />
+                  <AppWrapper wrapTokenPage={wrapTokenPage}>
+                    <ChainbridgeRoutes wrapTokenPage={wrapTokenPage} />
                   </AppWrapper>
                 </Router>
               </ChainbridgeProvider>