+++ src/Layouts/AppHeader.tsx
@@ -4,57 +4,82 @@ import clsx from "clsx";
 import { NavLink, Typography } from "@chainsafe/common-components";
 import { shortenAddress } from "../Utils/Helpers";
 import { useChainbridge } from "../Contexts/ChainbridgeContext";
-import CereLogo from "../media/Icons/cere-icon.svg";
+import CereLogo from "../media/Icons/bridge-icon.png";
 
-const useStyles = makeStyles(({ constants, palette, zIndex }: ITheme) => {
-  return createStyles({
-    root: {
-      display: "flex",
-      position: "fixed",
-      justifyContent: "space-between",
-      padding: `${constants.generalUnit * 2}px ${constants.generalUnit * 4}px`,
-      width: "100%",
-      top: 0,
-      left: 0,
-      backgroundColor: palette.additional["header"][1],
-      borderBottom: `1px solid ${palette.additional["header"][3]}`,
-      color: palette.additional["header"][2],
-      alignItems: "center",
-      zIndex: zIndex?.layer2,
-    },
-    left: {
-      display: "flex",
-      flexDirection: "row",
-      justifyContent: "flex-start",
-      alignItems: "center",
-    },
-    logo: {
-      height: constants.generalUnit * 5,
-      width: constants.generalUnit * 5,
-      "& svg, & img": {
-        maxHeight: "100%",
-        maxWidth: "100%",
+const useStyles = makeStyles(
+  ({ constants, palette, zIndex, breakpoints }: ITheme) => {
+    return createStyles({
+      root: {
+        display: "flex",
+        position: "fixed",
+        justifyContent: "space-between",
+        padding: `${constants.generalUnit * 2}px ${
+          constants.generalUnit * 4
+        }px`,
+        width: "100%",
+        top: 0,
+        left: 0,
+        backgroundColor: palette.additional["header"][1],
+        borderBottom: `1px solid ${palette.additional["header"][3]}`,
+        color: palette.additional["header"][2],
+        alignItems: "center",
+        zIndex: zIndex?.layer2,
+        [breakpoints.down("sm")]: {
+          flexDirection: "column",
+        },
       },
-    },
-    state: {
-      display: "flex",
-      flexDirection: "row",
-      alignItems: "center",
-    },
-    indicator: {
-      display: "block",
-      height: 10,
-      width: 10,
-      borderRadius: "50%",
-      backgroundColor: palette.additional["green"][6],
-      marginRight: constants.generalUnit,
-    },
-    address: {
-      marginRight: constants.generalUnit,
-    },
-    network: {},
-  });
-});
+      left: {
+        display: "flex",
+        flexDirection: "row",
+        justifyContent: "flex-start",
+        alignItems: "center",
+      },
+      logo: {
+        height: constants.generalUnit * 5,
+        width: constants.generalUnit * 5,
+        "& svg, & img": {
+          maxHeight: "100%",
+          maxWidth: "100%",
+        },
+      },
+      state: {
+        display: "flex",
+        flexDirection: "row",
+        alignItems: "center",
+      },
+      indicator: {
+        display: "block",
+        height: 10,
+        width: 10,
+        borderRadius: "50%",
+        backgroundColor: palette.additional["green"][6],
+        marginRight: constants.generalUnit,
+      },
+      address: {
+        marginRight: constants.generalUnit,
+      },
+      network: {},
+      accountInfo: {
+        display: "flex",
+        flexDirection: "row",
+        alignItems: "center",
+        justifyContent: "center",
+      },
+      mainInfo: {
+        display: "flex",
+      },
+      rightWrapper: {
+        display: "flex",
+      },
+      link: {
+        paddingRight: 10,
+      },
+      title: {
+        paddingLeft: 10,
+      },
+    });
+  }
+);
 
 interface IAppHeader {}
 
@@ -70,7 +95,11 @@ const AppHeader: React.FC<IAppHeader> = () => {
             <img src={CereLogo} alt="Cere Logo" />
           </NavLink>
         </div>
-        <NavLink style={{ textDecoration: "none" }} to="/">
+        <NavLink
+          style={{ textDecoration: "none" }}
+          className={classes.title}
+          to="/"
+        >
           <Typography variant="h4">Bridge</Typography>
           <br />
           <Typography variant="body1">
@@ -78,21 +107,41 @@ const AppHeader: React.FC<IAppHeader> = () => {
           </Typography>
         </NavLink>
       </div>
-      <section className={classes.state}>
-        {!isReady ? (
-          <Typography variant="h5">No wallet connected</Typography>
-        ) : (
-          <>
-            <div className={classes.indicator}></div>
-            <Typography variant="h5" className={classes.address}>
-              {address && shortenAddress(address)}
-            </Typography>
-            <Typography variant="h5" className={classes.address}>
-              connected to <strong>{homeConfig?.name}</strong>
-            </Typography>
-          </>
-        )}
-      </section>
+      <div className={classes.rightWrapper}>
+        <section className={classes.link}>
+          <a
+            href="https://cere-network.gitbook.io/cere-network/"
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            <Typography variant="h5">DDC Wiki</Typography>
+          </a>
+        </section>
+        <section className={classes.state}>
+          {!isReady ? (
+            <Typography variant="h5">No wallet connected</Typography>
+          ) : (
+            <>
+              <div className={classes.mainInfo}>
+                <div className={classes.accountInfo}>
+                  <span className={classes.indicator} />
+                  <Typography variant="h5" className={classes.address}>
+                    {address && shortenAddress(address)}
+                  </Typography>
+                </div>
+                <Typography variant="h5" className={classes.address}>
+                  <div>
+                    <span>connected to </span>
+                    <span>
+                      <strong>{homeConfig?.name}</strong>
+                    </span>
+                  </div>
+                </Typography>
+              </div>
+            </>
+          )}
+        </section>
+      </div>
     </header>
   );
 };