+++ src/types/globals.ts
@@ -1,0 +1,10 @@
+import { ChainbridgeConfig, UIConfig } from "../chainbridgeConfig";
+
+declare global {
+  interface Window {
+    __RUNTIME_CONFIG__: {
+      UI: UIConfig;
+      CHAINBRIDGE: ChainbridgeConfig;
+    };
+  }
+}