+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 - ...
 
+## v0.5.0
+
+- Added cross link on header
+- Updated favicon and icon
+- Updated with ChainSafe/chainbridge-ui v0.0.3
+
 ## v0.4.1
 
 - Renamed Cere Networks